вынесены функции из reviews и form
