+++ src/form.js
@@ -1,6 +1,7 @@
 'use strict';
 
 (function() {
+  var utils = require('./utilities');
   var formContainer = document.querySelector('.overlay-container');
   var formOpenButton = document.querySelector('.reviews-controls-new');
   var formCloseButton = document.querySelector('.review-form-close');
@@ -9,13 +10,9 @@
   var radio = form.elements['review-mark'];
   var name = form.elements['review-name'];
   var text = form.elements['review-text'];
-  var textPrompt = document.querySelector('.review-fields-text');
-  var namePrompt = document.querySelector('.review-fields-name');
-  var formPrompt = document.querySelector('.review-fields');
   var reviewButton = document.querySelector('.review-submit');
   var RADIO_MIDDLE_VALUE = 3;
-  var BIRTH_DAY = 19;
-  var BIRTH_MONTH = 11;
+
   var cookies = require('browser-cookies');
 
   document.getElementById('review-name').required = true;
@@ -25,7 +22,7 @@
   name.value = cookies.get('nameValue');
 
   form.oninput = function() {
-    changeVisible();
+    utils.changeVisible(name.value, text.value);
     controlButton();
   };
 
@@ -37,44 +34,9 @@
     }
   };
 
-  var changeVisible = function() {
-    if (name.value) {
-      namePrompt.classList.add('invisible');
-    } else {
-      namePrompt.classList.remove('invisible');
-    }
-
-    if (text.value) {
-      textPrompt.classList.add('invisible');
-    } else {
-      textPrompt.classList.remove('invisible');
-    }
-
-    if (name.value && text.value) {
-      formPrompt.classList.add('invisible');
-    } else {
-      formPrompt.classList.remove('invisible');
-    }
-  };
-
-  var daysToExpire = function() {
-    var now = new Date();
-    var myBirthday = new Date();
-    myBirthday.setFullYear(now.getFullYear(), BIRTH_MONTH, BIRTH_DAY);
-
-    var difference = now - myBirthday;
-
-    if (difference < 0) {
-      myBirthday.setFullYear(now.getFullYear() - 1, BIRTH_MONTH, BIRTH_DAY);
-      difference = now - myBirthday;
-    }
-
-    return difference;
-  };
-
   form.onsubmit = function() {
-    cookies.set('radioValue', radio.value, {expires: Date.now + daysToExpire()});
-    cookies.set('nameValue', name.value, {expires: Date.now + daysToExpire()});
+    cookies.set('radioValue', radio.value, {expires: Date.now + utils.daysToExpire()});
+    cookies.set('nameValue', name.value, {expires: Date.now + utils.daysToExpire()});
   };
 
   field.onchange = function() {
